requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
pdfplumber==0.9.0
python-docx==0.8.11